pyproject.toml
[project]
name = "video-subtitle-generator"
version = "0.1.0"
description = "Video Subtitle Generator"
requires-python = ">=3.9,<3.13"
dependencies = [
    "openai-whisper",
    "torch",
    "torchaudio",
    "torchvision",
    "ffmpeg-python",
]

[tool.uv]
dev-dependencies = ["pytest", "black", "flake8", "mypy"]
